+++ static/python/frontend/weather_frontend_api.py
