+++ static/python/frontend/weather_report.py
