+++ templates/index.html
@@ -24,9 +24,9 @@
     </py-config>
     <py-env>
         - paths:
-            - /static/python/client.py
-            - /static/python/weather_frontend_api.py
-            - /static/python/weather_report.py
+            - /static/python/frontend/client.py
+            - /static/python/frontend/weather_frontend_api.py
+            - /static/python/frontend/weather_report.py
     </py-env>
 </head>
 <body id="the_body">
@@ -161,7 +161,7 @@
         A pyscript demo app originally created by <a href="https://twitter.com/mkennedy" target="_blank">@mkennedy</a>, copyright 2022 (Modified by Marcus Holmström).
     </div>
 </footer>
-<py-script src="/static/python/client.py"></py-script>
+<py-script src="/static/python/frontend/client.py"></py-script>
 <script src="/static/js/pwa-scaffold.js"></script>
 </body>
 </html>
